+++ dchars/languages/hbo/dstring.py
@@ -150,6 +150,11 @@ class DStringHBO(DStringMotherClass):
 
                 NB : function required by the dchars-fe project.
         """
+
+        # Pylint can't know that <self> has a 'transliteration_method' member
+        # created when <self> has been initialized by new_dstring() :
+        # pylint: disable=E1101
+        # -> "Instance of 'DStringHBO' has no 'transliteration_method' member"
         res = DStringHBO.trans__get_transl_ucombinations[self.transliteration_method]()
         return res