+++ dchars/languages/lat/dstring.py
@@ -147,6 +147,11 @@ class DStringLAT(DStringMotherClass):
 
                 NB : function required by the dchars-fe project.
         """
+
+        # Pylint can't know that <self> has a 'transliteration_method' member
+        # created when <self> has been initialized by new_dstring() :
+        # pylint: disable=E1101
+        # -> "Instance of 'DStringLAT' has no 'transliteration_method' member"
         res = DStringLAT.trans__get_transl_ucombinations[self.transliteration_method]()
         return res