+++ dchars/languages/san/dstring.py
@@ -157,6 +157,11 @@ class DStringSAN(DStringMotherClass):
 
                 NB : function required by the dchars-fe project.
         """
+
+        # Pylint can't know that <self> has a 'transliteration_method' member
+        # created when <self> has been initialized by new_dstring() :
+        # pylint: disable=E1101
+        # -> "Instance of 'DStringSAN' has no 'transliteration_method' member"
         res = DStringSAN.trans__get_transl_ucombinations[self.transliteration_method]()
         return res